# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Alga Taxi API (scaffold)
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Local dev
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OrderPoint:
      type: object
      properties:
        type:
          type: string
          enum: [pickup, dropoff]
        address:
        AppSettings:
          type: object
          properties:
            show_phone_numbers:
              type: object
              properties:
                enabled:
                  type: boolean
            contact_number:
              type: string
            logo_url:
              type: string
          type: string
        lat:
          type: number
        lon:
          type: number
    Order:
      type: object
      properties:
        id:
          type: string
        client_id:
          type: string
        driver_id:
          type: string
        status:
          type: string
        service_type:
          type: string
        price_cents:
          type: integer
        points:
          type: array
          items:
            $ref: '#/components/schemas/OrderPoint'
...
security:
  - bearerAuth: []
paths:
...
  /api/orders:
    post:
      summary: Create order (client)
      tags: [orders]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                service_type:
                  type: string
                points:
                  type: array
                  items:
                    $ref: '#/components/schemas/OrderPoint'
                payment_method:
                  type: string
                notes:
                  type: string
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'

    get:
      summary: List orders (user-specific)
      tags: [orders]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
      responses:
        '200':
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
...
  /api/admin/commission:
    get:
      summary: List commission rules
      tags: [admin]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Commission rules
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object

    post:
      summary: Create commission rule
      tags: [admin]
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommissionRule'
      responses:
        '201':
          description: Created
  /api/admin/commission/{id}:
    patch:
      summary: Update commission rule
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommissionRule'
      responses:
        '200':
          description: Updated

    delete:
      summary: Delete commission rule
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
  /api/admin/blocks:
    post:
      summary: Create manual block for a user
      tags: [admin]
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                reason:
                  type: string
                blocked_until:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Created
  /api/admin/blocks/{id}:
    delete:
      summary: Remove block (unblock)
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Removed
...
components:
  schemas:
    CommissionRule:
      type: object
      properties:
        id:
          type: string
        scope:
          type: string
        service_type:
          type: string
        driver_id:
          type: string
        percent:
          type: number
...